Renamed restfulptv test package to ptv
